Misplaced vertical bar line in append_tabbed
Running the example below results in vertical bar to be offset too far
to the right by one position.

. `dirname $0`/simple_curses.sh

main(){
  window "Test Window" "white" "20%"
  append_tabbed "One:Two:Three" 3
  endwin
}
main_loop 5

The vertical bar is shifted by one position to the right for the tabbed
row. This patch compensates for one shift. The last column position for
the text is computed from window last column (including padding of one
position) minus vertical bar (one position) and one position for a space
between vertical bar and last character. This avoid the text to stick to
vertical bar.