Color settings for cells applied to borders
When using append_tabbed with color settings, the color is kept longer
than necessary. As consequence the color of the last element is used as
well for the window right border, and the border of next line. In the
worst case, the whole last horizontal line is drawn using the color
setting for the last tabbed element.

This change resets color to defaults (system) for drawing windows
borders.