Improve fallback randomness for temporary file
I didn't want to bring in yet more programs and pipes, but I think it's
worth it, requiring only _coreutils_ and access to '/dev/urandom'. I'm
not quite sure how portable that device is, though.

Some people online claim they have issues with tr(1) and this method on
Mac. It's apparently solvable by redirecting directly into tr(1). I see
no mention of which OS this project is for, so I shall assume Linux.

The approach I took _is_ actually doable in BASH:

  (
      LC_ALL=C

      Chars=
      while (( ${#Chars} < 10 )); do
          read -n 1 < /dev/urandom
          [[ $REPLY == [[:alnum:]] ]] && Chars+=$REPLY
      done

      printf '%s\n' "$Chars"
  )

Nice thing about that is that it keeps it in BASH, doing away with
tr(1) and head(1). It works by repeatedly grabbing a single byte from
'/dev/urandom', then checking to see if it's good. While the variable
is less than the length we want, it'll keep looking for characters.
It's very quick, contrary to what you might expect.

I'm just not sure how portable and reliable it is. Works well for
_me_, but of course that's a very small data set.

I quote-protected `$rand` on the very off-chance something funky happen
that BASH decided to try to interpret, BTW.