Optimize and untangle getting length of strings
The previous use of process substitution with command substitution, -
wc(1), and a call to the `echo` builtin was quite inefficient, although
I can certainly appreciate the out-of-the-box thinking. I strongly
recommend that you look at bash(1) for PARAMETER EXPANSION, as there
are a _lot_ of amazing features right there in the shell.

I've also combined the assignment and declaration to reduce noise. The
`local` builtin actually lets you assign and declare multiple times
with a single `local` call, which winds up being more efficient.

This is actually another example of inconsistency causing confusion.
Parameter expansion is used elsewhere to achieve this, yet in other
places the old approach is used, so I can only _hope_ there wasn't a
specific reason.